DOCD-1243 Use port name if single tcp port with no explicit health check
It think this has never worked; if an application has a single tcp port specified but no explicitly set health check,
the first and only port is used in a tcpSocket check as both the liveness and readiness check. The issue is that we
use the port field from the lookup, which is the external Service port, not the target_port which is the container
port. This means that the health check would always fail unless these fields happened to have the same value.
port.

Change this behavior to use the name field to mimic the behavior for the http check.