Reduce the amount of testcases for deployment deployer
We were generating one test case for every combination of values from all fixtures for these tests. I tried to add a few parameters to these tests when developing a new feature and ended up with > 20k test cases. It is not neccessary to test every combination of everything, so I've rewritten the tests to use config and app_spec as fixtures, and thus reduced the number of tests