Add support for stripping defaults when doing transformation
This allows for returning transformed configs without the defaults that
may otherwise be applied by the transformer(s).

For lists where there is a single default item of collection.Mapping defined
(ie 'ports') the configuration will only be stripped if it contains the default
or a subset of the default collection.mapping as its only item.