Always unzip eggs / workaround setuptools#1256
It seems that the file system race that happens in CI occasionally is an edge case of an edge case. If a dependency
is installed as a zipped egg, and imported at the relative same time by multiple threads, the error may be
triggered. This workaround is to tell setuptools to always unzip eggs when installing them.

See also https://github.com/pypa/setuptools/issues/1256#issuecomment-360216975