Status cleanup (#191)
When creating or updating a status-object, delete old ones for the same application. Currently we keep the last 10.
The first time this runs, we don't know which are actually the last, so we keep 10 more or less randomly selected statuses. This might be worth investing time into avoiding, I'm not sure. It can also be fixed by manually applying the `last_updated` annotation to the ones we want to keep before merging this.


Squashed commits of the pull request:
* Add last-updated annotation for sorting later


* Delete all statuses except the last 10, ordered by last-updated


* Use UTC for times


* Add utility-script used for cleanup


* Merge master into status-cleanup