Add "singleton" support and avoid scaling down when rolling updates
If replicas.singleton is set to true and replicas.maximum is 1, there
will be downtime during deployment as there is only a single instance
running at any one time.  If replicas.singleton is false, another
instance will first be started.

Also change the defaults for RollingUpdate strategy to be maxSurge: 1,
maxUnavailable: 0 to prevent losing too much capacity for services
with few instances.