Add support for ingress tls
This introduces functionality to extend ingress objects with tls when
they are created. This can be configured to be added per application as well
as to be added by default.

A configuration flag `--use-ingress-tls` is used to control the desired behavior
`default_on`, `default_off` or `disabled`.

By default this functionality is disabled.