Use pipeline annotation fallback to determine repository
In order to be able to provide source repository information as part of usage reporting
the usage reporter relies on the fiaas/source-repository annotation for the deployment
to contain a value. Only recently have we started to populate this value as part of our
base pipelines so anything that has not been deployed after those changes took place will
not have the annotation populated. There is however another annotation that has been
around for longer that we could use as fallback to be able to include this information.

This fallback can then be removed eventually.