Add trigger type info
Devhose requires the trigger type to be set to be able to process the events.
Currently this is simply set to a default that is sufficient but could be extended
in future to provide some additional context information that could be used
to tie the deployment back to something that would have triggered it such as
a build system.