Use same value for infrastructure and provider
Due to the way devhose processes events both of these values need to be set
but in the case of FIAAS they contain the same information so we just duplicate
the data to fulfil the contract with devhose.
Previously provider contained the team name but they have opted to change that
and dedicate a separate field for it.