Provide a default team for usage reporting events
Devhose tries to determine the team/organization that the usage
event is being done for by inspecting the values contained within
the usage event. In some cases the usage event may not contain enough
information for that to be done successfully. In that case we need
to provide a default value that they can use.

This adds adds the 'team' field and populates it with the operator
identifier.