Add config flag for specifying usage reporting team
This is an optional flag but can be useful to start with until some of the
repository details are included in the usage events. It will allow us to provide
a default based on the environment such as namespace to provide some value that will
allow more granularity in the usage metrics.