Add flag to make emptyDirs use memory medium (#238)
The SCHIP cluster has problems with nodes getting their FS mounted read-only under some circumstances, and the CRE team has found a big that links this to the use of emptyDir volumes.

In FIAAS, applications get their `/tmp` as an emptyDir, and also secrets are mounted in an emptyDir if the secret init container solution is in use.

To help the CRE team, we implement a flag, which will change the emtpyDirs to be backed by memory, which should solve the problem. It needs to be a flag, because the memory used is considered part of the container limits, and could potentially cause problems in other clusters where use of `/tmp` is more widespread.

We will enable this flag only in the SCHIP cluster.