Signal deployment failure when unable to parse fiaas config
We currently consider a deployment started when an app_spec is picked up by
the deployer thread. This isn't strictly speaking correct, as there are things
happening before this that would be natural to consider part of the deployment
(most notably, parsing of fiaas.yml).  The most visible problem with this
is that for configuration errors we don't create a status object at all.

This introduces error handling outside of the deployer that will signal failure
when unable to parse fiaas configuration for a given deployment.
This will result in a status resource to be created for the deployment and a
usage reporting event to be sent.