Remove fallback to operator when team is not set for usage reporting
With event processing being extended with additional mechanisms to determine
team from deployment target infrastructure in many cases we can reduce the
importance on supplying team information in deployment events. Furthermore to
remove the fallback to operator identifier as that may not necessarily be
accurate and instead only communicate the value when explicitly set.