Retry upsert operation on any Kubernetes resource on 409 Conflict
We are seeing occasional `409 Conflict` responses when updating Deployment resources. I'm not 100% sure why this
happens, but I suspect that it has to do with a delay in updating the `resourceVersion` of a resource internally
within Kubernetes. When deployments of an application occur quickly enough after one another, we can get into a
situation where the `resourceVersion` we got by GET-ing the resource might be updated on the server side before we get
around persisting our changes to the resource by POST- or PUT-ing the modified resource. In this situation it is
recommended to GET the resource again, apply changes before POST/PUT-ing again[1], so this is what I'm implementing in
this change.

I'm also adding counters `fiaas_upsert_conflict_retry` which counts the number of retries for a given `target` (this
is a label on the counter) function and fiaas_upsert_conflict_failure which counts the number of times we exceeds the
number of retries. We should be able to use these to determine exactly how often this behavior occurs.

[1]: https://github.com/kubernetes/community/blob/master/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency