Separate endpoints for apiserver when connecting from container and host
We are running Kind in a container, and need to be able to connect to its apiserver from fiaas-deploy-daemon either
running in a container or on the host, and from the tests themselves, which always runs on the host. Kind exposes the
apiserver on port 8443 in the container it runs, and we also map this to `localhost:$RANDOM_PORT` on the host that
runs the container.

On Linux, we can run Docker with host networking. This runs all containers on the hosts network, meaning we can
connect to ports bound on localhost whether we're running inside a container (fiaas-deploy-daemon) or not (the tests
themselves), and it will work fine. On macOS with Docker for mac, it is not possible to use host networking, so we
have to use different endpoints when connecting to the apiserver running in one container from the host and from
fiaas-deploy-daemon running in another container. This commit changes the kubernetes['server'] endpoint into
kubernetes['container-to-container-server'] for connecting from fiaas-deploy-daemon running in a container, and
kuberentes['host-to-container-server'] for connecting from fiaas-deploy-daemon running on the host, and for connecting
from the tests themselves.

The TLS certificate it uses is pre-generated, and has `localhost` and `kubernetes` among others as valid SANs. To make
sure we can validate the cert when connecting from fiaas-deploy-daemon in a container, we use `--add-host` to map
`kubernetes` to the Kind container's IP in the fiaas-deploy-daemon container's hosts file.