Fix scale up issue when autoscaler is enabled and application downscaled
This fix aims to set the number of replicas to replicas.maximum when the
current deployment has 0 replica set. Otherwise, the number of replicas stays
to 0 and the application is never upscaled (the HPA is disabled with this
state: "scaling is disabled since the replica count of the target is zero").