Refactor lifecycle signals
The list of parameters to the signals has grown quite long, and each
received the same set of parameters, and the handlers essentially handled
each in the same way. This switches to use a single signal, and replaces
the list of parameters with a 'Subject' that wraps them, along with a status.

Fixes #34