Reduce coupling to code under test in test_action_on_signal
Move mocking one layer down, from the model layer to the api call layer. This
makes the test less coupled to the behavior in the production code. It
reduces the complexity of mocking the state of resources, since the test can
mock what is returned from the api without knowing if get or get_or_create is
used in the function under test.