Allow datadog container memory to be set via config
We're seeing that the sidecar is using a lot less than 2Gi of memory,
(with a maximum of 220Mi)
and across the cluster this adds up to a lot of reserved, unused resources.
This makes the amount configurable, so that we can experiment with what
a reasonable value will be.

Fixes #75