Refactor init-container secrets
This rewrites the processing of the default init-container,
and the strongbox version, to be based on the same implementation
that will be re-usable for the new functionality proposed in #71.

The initialisation of an empty list of `secret_specs` in Secrets.apply will be
replaced by the secrets specified in the app-spec, and the map of
`_available_secrets_containers` in GenericInitSecrets will be replaced
by a value taken from the deploy-daemon config.