Support multiple secret-init-containers
Fixes #71

This change allows FIAAS operators to register multiple
images that can be used by app-developers as init-containers
for pulling secrets that are made available to the app.

Operators provide a mapping between a 'type' and an image, e.g.
`aws-parameter-store=fiaas/parameter-store-secrets`
And then app developers specify that they wish to use this in their
paas.yml:
```yaml
extensions:
  secrets:
    aws-parameter-store:
      parameters:
        AWS_REGION: eu-west-1
        SECRET_NAME: somesecret
      annotations:
        iam.amazonaws.com/role: arn:123:role/some-role
```

The 'parameters' will be set as ENV vars in the container, while
'annotations' will be added to the pod.

If an operator registers an image with 'type' set to 'default',
then this image will be attached automatically, iff the app config
doesn't specify anything.

Include a new e2e test for a registered secret init-container

Documentation, in the v3 spec, for the new feature.

Add operator documentation too