Make deployments default to minimum replicas
FIAAS sets deployment replicas to max (from the resources definition) by
default. This is done with the expectation that tha HPA will reduce the
number of pods to the minimum/correct number with time. However, for
those environments where there's no HPA because no enforcement is done
(in our case, all dev envs and most pre envs) that never happens. This leaves apps
with an unnecessary high number of pods.