Create stable names for ApplicationStatuses
create_name used the `hash()` builtin, which has changed between python2 and python3 in two ways:
- in python3 , hash uses a different hash algorithmm
- hashes aren't stable between python processes because the hash
function is seeded with a random value on interpreter startup

This changes create_name to use a hash function that produces stable
hashes. I'm not sure if it is the best algorithm for the purpose, but I
chose it because it is one of the algorithms which is always available
in hashlib, and it can be configured to output a variable length
hash. We don't want a super long hash because the hash is concatenated
with the application name and that has to fit within 253 characters.

Since the hash function had to change when switching to python 3, the
logic that checks for an existing ApplicationStatus might have to be
changed to look up the ApplicationStatus by fiaas/deployment_id label
and not name to avoid redeploying already deployed applications that
were deployed with a python2 fiaas-deploy-daemon (I've added a todo
comment in the relevant place)