Remove code-path for responding to Delete events
Now that we are setting ownerReferences on the created objects
kubernetes will take care of deleting them for us when the application
is deleted, so we can remove all the code related to that.

The delete methods for service, ingress and autoscaler are still needed
for the case where the config changes such that they are no longer
required, but now they can be private/internal methods.