Disable setting config options via environment variables
fiaas-deploy-daemon's configuration is based on a merge of command-line
flags, environment variables and the cluster_config.yaml configuration
file, typically provided via a ConfigMap.
Because of the order of precedence (command-line flags > environment
variables > config file), and because fiaas-deploy-daemon sets
environment variables from its configuration when deploying itself, this
could lead to situations where a change in the config file would not
take effect if the configuration option that was change was shadowed by
a environment variable.

To remove the potential for confusion stemming from this behavior,
remove the ability to configure fiaas-deploy-daemon via environment
variables.