Separate hosts into ingresses when they have different issuer types
This changes the grouping of hosts into separate ingresses, previously
doing so based on annotations, to additionally create separate ingress(es)
for different overridden issuer-types. This allows for a domain-suffix
to be registered as requiring, say, certmanager.io/issuer (instead of the
default, /cluster-issuer), and an extra ingress will be created, and
it will be annotated approprately.

This closes #114