log warning if config env variable exists
There was a bug that caused fiaas-deploy-daemon to raise an exception on
startup if host rewrite rules were configured, because of the logic in
warn_if_env_variable_config which tries to convert from config flag to
config environment variable value.

I don't think it is actually necessary or useful to check the value of
the environment variable in this context, so now
warn_if_env_variable_config just checks if the environment variable keys
are present in the environment.

Many of the config environment variable keys which were checked for were
actually wrong too, I've fixed those too