Keep autoscaler min_replicas=max_replicas if max < min
If the user specifies a minimum number of replicas higher than the
max number of replicas, the autoscaling will set the deployment
number of replicas to the minimum (which is higher than the maximum).

Setting max < min doesn't make much sense in any case, but this
is even more unexpected when the user specify a max=1 and doesn't
even specify minimum and it takes the default(2).

This fix makes sure autoscaler.min_replicas is never higher than
autoscaler.max_replicas.

Signed-off-by: Xavi León <xavi.leon@adevinta.com>