fix: Ensure the default host has a path defined
Fix the case when the default ingress is wrongly created when only annotated
hosts are specified. For example, a fiaas spec like this one:

application: myapp
config:
    ingress:
    - host: myhost
      paths:
      - path: /
      annotations:
        kubernetes.io/ingress.class: internal

leads to the create the default ingress as follow:

spec:
  rules:
  - host: myapp.192.168.49.2.xip.io

This commit ensure that the default ingress is always created with the default
path.