Don't get non-existing ApplicationStatus
Using get_or_create here with generateName (and implicitly name=None)
results in a GET to .../application-statuses/None, which is unnecessary,
and could theoretically result in writing to an appstatus named `None`
if one exists.

If we go into this branch, there is no ApplicationStatus for this
app_name,deployment_id, so it should be safe to skip get_or_create and
just have it be created when .save() is called below.