Use k8s.config.use_in_cluster_config for in-cluster configuration
This is mainly to support regularly re-reading the service account token
from file when in-cluster configuration is used.

This also introduces the option to explicitly override the default
in-cluster configuration by setting the api-token or api-cert config
flags. Currently these flags are ignored if
`/var/run/secrets/kubernetes.io/serviceaccount/token` exists, which
could be surprising behaviour.

If neither api-token or client-cert is set, and in-cluster configuration
can not be set up (service account token can not be read), log a warning
indicating potentially missing apiserver authentication configuration.