Keep Datadog sidecar alive while main application is too
While the main application is given a grace period to terminate, and has
the termination signal delayed using a preStop action, the datadog
sidecar would be immediately terminated. This would lead to a loss of
metrics from before the pod received the termination signal, and to the
point it actually exited.

With this change, the application's pre-stop delay is extended to the
sidecar, with an additional 5 seconds to give the main container a
chance to exit gracefully while metrics are still being collected. No
guarantees, regarding metrics collection, are offered to applications
that take longer to exit.