Upgrade monotonic; fix scheduler logic on Apple arm64
Upgrade monotonic to 1.6, which includes a fix for monotonic.monotonic
not correctly returning fractional seconds on Apple arm64[1]. This fixes
an issue where when running fiaas-deploy-daemon on Apple arm64, the
Scheduler thread would not execute ReadyChecks when it was supposed
to (never?). In practice, this fixes an issue with running the bootstrap
e2e tests on Apple silicon, as these tests verify that deployments
complete successfully (i.e. waits for ReadyCheck).

[1]: https://github.com/atdt/monotonic/releases/tag/1.6