Temporarily disable html test report
When html test reports are enabled, the summary "n tests Passed"
indicates twice the number of tests which is actually run. I think this
might be because a bug or compatibility issue problem with pytest-html
3.x, because if html test reports are disabled, the summary test count
is correct.

In the output from pytest-xdist, and in the generated coverage and
actual html reports (when enabled) the number of tests are reported
correctly in both case.

It seems like this issue is fixed in the most recent version of
pytest-html, but upgrading to that requires also upgrading flask and
associated dependecies to the next major version, and I'd prefer to do
that as a separate changeset.