Update flask and related dependencies
Flask and its transitive dependencies were pinned to old versions which
supported python 2. Upgrade to the most recent versions, since python 2
support is no longer relevant.

html test reports were disabled in
9f42cd9cebaf2aeed8a54aafc3bb19bfda99b383 because of an interaction issue
with the older version of pytest-html which was used. Since flask and
jinja2 is updated now, it is possible to use the most recent version of
pytest-html, which does not have the duplicate test number reporting
issue. Update pytest-html and and re-enable html reports for tests.