Handle exception from lifecycle.failed in deployer error handler
lifecycle.failed calls the Kubernetes API to update Application and
ApplicationStatus resources. If any of these fail with e.g. a
ClientError, the deployer thread will exit and fiaas-deploy-daemon will
be restarted because the liveness probe fails.
If this failure mode is triggered by a non-recoverable error, such as
the resource is too large, this can cause disruption to deploys, since
fiaas-deploy-daemon would restart every time it tries to handle the
Application triggering the error. Log and drop exceptions from
lifecycle.failed to avoid this situation.