Handle exception from task (ReadyCheck)
ReadyChecks calls the Kubernetes API to check Deployment status, and to
update Application and ApplicationStatus resources. If any of these
requests fail with e.g. a ClientError, the scheduler thread will exit
and fiaas-deploy-daemon will be restarted because the liveness probe
fails. If this failure mode is triggered by a non-recoverable error,
such as the resource is too large, this can cause disruption to deploys,
since fiaas-deploy-daemon would restart every time it tries to handle
the Application triggering the error. Log and drop exceptions from
tasks to avoid this situation. The failed task is also dropped, so
deployment will be stuck with RUNNING status, and should be redeployed
if fiaas-deploy-daemon restarts.