Define project and dependencies to pyproject.toml
The primary motivation for this is to get the Github dependency graph to
resolve properly so that dependabot security alerts can work in this
repository. Apparently the system that resolves the dependency graph
can be unreliable when dependencies are defined in setup.py[1], and
using variables when specifying dependencies may also cause issues[2].

pyproject.toml / PEP621 seems like the new standard of configuring
python projects and associated tooling, so I think this is a step in the
right direction in any case.

A minor difference is that the `[dev]` dependency target doesn't include
the dependencies from the `[ci]` target. To work around that, it is
necessary to use `[dev,ci]` to achieve the same thing as `[dev]`
previously. requirements.txt is updated to reflect this, so
`pip install -r requirements.txt` should still install all dependencies.

[1]: https://docs.github.com/en/code-security/supply-chain-security/understanding-your-software-supply-chain/about-the-dependency-graph#supported-package-ecosystems
[2]: https://docs.github.com/en/code-security/supply-chain-security/understanding-your-software-supply-chain/troubleshooting-the-dependency-graph#does-the-dependency-graph-detect-dependencies-specified-using-variables