Use same behaviour as with setup.py for finding packages
When using find_packages via pyproject.toml, namespaced packages are
automatically included. This means all directories in the repository
root is included by default and we don't want that. The ideal solution
might be to change the repo structure to src-layout[1], but for now use
namespaces=false to use the same behaviour with setup.py find_packages
to keep the changeset as small as possible.

[1]: https://setuptools.pypa.io/en/latest/userguide/package_discovery.html#src-layout