Skip updates on Application with deletionTimestamp in the past
metadata.deletionTimestamp is set when a resource is deleted with
propagationPolicy: orphan or propagationPolicy: foreground. The resource
may be updated after this, for example finalizers may be set or removed.
If the application resource is in the process of being deleted (i.e. has
deletionTimestamp set and in the past), skip deploy deploy to avoid
interfering with the deletion process done by the garbage collector.