Add an ingress-host-pair option.
If ingresses exists on one of the hosts in the pair, they're also
created on the other one, provided there's not already an ingress
defined for that host.

This is done after rewrite rules, it thus applies to the rewritten
hosts.

Motivation for this change is to easily change the main ingress host of
a large number of applications.